Prevent past appointment cancellation

+++ app/controllers/Databases/AppointmentsDB.java
@@ -325,8 +325,10 @@ public class AppointmentsDB {
 
     public static AppointmentsModel removeAppointment(String appointmentId){
         AppointmentsModel appointment = getAppointment(appointmentId);
+        if(!appointment.getStartDate().before(new Date())){
+            ApiFuture<WriteResult> writeResult = FirestoreDB.get().collection("appointments").document(appointmentId).delete();
+        }
         /* Asynchronously remove appointment from DB */
-        ApiFuture<WriteResult> writeResult = FirestoreDB.get().collection("appointments").document(appointmentId).delete();
         return appointment;
     }
 }
